modification navbar : suppression du composant pour travail sur les link
