+++ frontend/src/App.jsx
@@ -1,6 +1,6 @@
 import { BrowserRouter } from "react-router-dom";
 import "./styles/App.scss";
-import Navbar from "./components/Navbar";
+import Navbar from "./components/Navbarbis";
 import MainPage from "./components/MainPage";
 
 function App() {
