+++ frontend/src/styles/Navbar.scss
@@ -31,19 +31,7 @@
   margin-right: 0.5rem;
 }
 
-.isNotActiveOnglet {
-  display: none;
-}
-
 @media screen and (min-width: 767.98px) {
-  .navbarContainer {
-    flex-direction: column;
-    justify-content: flex-start;
-    align-items: center;
-    width: 4rem;
-    height: 100vh;
-    position: fixed;
-  }
 
   .navbar {
     flex-direction: column;
@@ -99,13 +87,30 @@
     cursor: pointer;
   }
 
-  .logo {
-    margin-left: 0;
-    margin-top: 1rem;
+
+
+  .iconeList {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-around;
+      img {
+        padding:0;
+      }
   }
 
-  .login {
-    margin-right: 0;
-    margin-bottom: 1rem;
+  .navbarContainer {
+    display:flex;
+    flex-direction: column;
+    justify-content: space-between;
+    align-items: center;
+    flex-direction: column;
+    width: 4rem;
+    height: 100vh;
+    position: fixed;
+      img{
+        padding:10px 0;
+      }
   }
+
+
 }
